ACTIN-170 Persist/read the config with the orange record
Persist the full input config with the orange record such that it can be read back and used
when rerunning.
