+++ orange/src/main/java/com/hartwig/hmftools/orange/OrangeApplication.java
@@ -2,6 +2,7 @@ package com.hartwig.hmftools.orange;
 
 import java.io.IOException;
 
+import com.hartwig.hmftools.datamodel.orange.OrangeConfig;
 import com.hartwig.hmftools.datamodel.orange.OrangeRecord;
 import com.hartwig.hmftools.orange.algo.OrangeAlgo;
 import com.hartwig.hmftools.orange.report.ReportWriter;
@@ -25,11 +26,11 @@ public class OrangeApplication {
     public static void main(String[] args) throws IOException {
         LOGGER.info("Running {} v{}", APPLICATION, VERSION);
 
-        Options options = OrangeConfig.createOptions();
+        Options options = OrangeCommandLine.createOptions();
 
         OrangeConfig config = null;
         try {
-            config = OrangeConfig.createConfig(new DefaultParser().parse(options, args));
+            config = OrangeCommandLine.createConfig(new DefaultParser().parse(options, args));
         } catch (ParseException exception) {
             LOGGER.warn(exception);
             new HelpFormatter().printHelp(APPLICATION, options);
